fix error with LS and SS

+++ client/src/app/data/services/auth.service.ts
@@ -61,7 +61,11 @@ export class AuthService {
         if (currentLink.includes('auth')) {
           this.router.navigate(['']);
         }
-        sessionStorage.clear();
+        if (typeof window !== 'undefined' && sessionStorage) {
+          if (typeof window !== 'undefined' && sessionStorage) {
+            sessionStorage.clear();
+          }
+        }
       }),
 
       catchError((error) => {
@@ -82,6 +86,8 @@ export class AuthService {
   logout() {
     this.cookieService.deleteAll();
     this.store.dispatch(authActions.logoutUser());
-    sessionStorage.clear();
+    if (typeof window !== 'undefined' && sessionStorage) {
+      sessionStorage.clear();
+    }
   }
 }
